Add structure to package and refactor modules
