+++ aieng-eval-agents/aieng/agent_evals/impl/__init__.py
@@ -1,6 +1,6 @@
 """Example implementation for an example topic."""
 
-from aieng.topic.impl.example_impl import example_impl
+from aieng.agent_evals.impl.example_impl import example_impl
 
 
 __all__ = ["example_impl"]
